+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/EndToEndTimeOutValidationTests.java
@@ -296,6 +296,44 @@ public class EndToEndTimeOutValidationTests extends TestSuiteBase {
     }
 
     @Test(groups = {"fast"}, timeOut = 10000L, retryAnalyzer = FlakyTestRetryAnalyzer.class)
+    public void queryChangeFeedWithEndToEndTimeoutPolicyAndAvailabilityStrategyShouldTimeout() {
+        if (getClientBuilder().buildConnectionPolicy().getConnectionMode() != ConnectionMode.DIRECT) {
+            throw new SkipException("Failure injection only supported for DIRECT mode");
+        }
+
+        CosmosAsyncClient cosmosClient = initializeClient(endToEndOperationLatencyPolicyConfig);
+        FaultInjectionRule faultInjectionRule = null;
+        try {
+            CosmosEndToEndOperationLatencyPolicyConfig endToEndOperationLatencyPolicyConfig =
+                new CosmosEndToEndOperationLatencyPolicyConfigBuilder(Duration.ofSeconds(1))
+                    .availabilityStrategy(
+                        new ThresholdBasedAvailabilityStrategy(
+                            Duration.ofMillis(100), Duration.ofMillis(200)))
+                    .build();
+
+            CosmosChangeFeedRequestOptions options =
+                CosmosChangeFeedRequestOptions.createForProcessingFromBeginning(FeedRange.forFullRange());
+            options.setCosmosEndToEndOperationLatencyPolicyConfig(endToEndOperationLatencyPolicyConfig);
+
+            faultInjectionRule = injectFailure(createdContainer, FaultInjectionOperationType.READ_FEED_ITEM, null);
+            CosmosPagedFlux<TestObject> changeFeedPagedFlux =
+                createdContainer.queryChangeFeed(options, TestObject.class);
+
+            StepVerifier.create(changeFeedPagedFlux)
+                        .expectErrorMatches(throwable -> throwable instanceof OperationCancelledException
+                            && ((OperationCancelledException) throwable).getSubStatusCode()
+                            == HttpConstants.SubStatusCodes.CLIENT_OPERATION_TIMEOUT)
+                        .verify();
+        } finally {
+            if (faultInjectionRule != null) {
+                faultInjectionRule.disable();
+            }
+
+            safeClose(cosmosClient);
+        }
+    }
+
+    @Test(groups = {"fast"}, timeOut = 10000L, retryAnalyzer = FlakyTestRetryAnalyzer.class)
     public void queryItemWithEndToEndTimeoutPolicyInOptionsShouldTimeoutWithClientConfig() {
         if (getClientBuilder().buildConnectionPolicy().getConnectionMode() != ConnectionMode.DIRECT) {
             throw new SkipException("Failure injection only supported for DIRECT mode");